+++ CHANGELOG.md
@@ -3,10 +3,8 @@
 # PEST IntelliJ Changelog
 
 ## Unreleased
-
-## 1.9.1-EAP.1 - 2023-02-09
-
 ### Fixed
+- Fixed ComposerLibraryManager being nullable now.
 - Fixed running tests with filenames containing `_`.
 
 ### Changed